requirements/requirements.txt
bitsandbytes==0.45.3; (platform_machine == 'x86_64' and platform_system != 'Darwin')
liger-kernel==0.5.8; (platform_machine == 'x86_64' and platform_system != 'Darwin')
cut-cross-entropy @ git+https://github.com/apple/ml-cross-entropy.git@87a86ab
torchdata
